基本完成数据交互VGDao层，未测试。 下午插入数据并进行简单测试。

+++ VG/src/com/ne/vg/bean/SmallScene.java
@@ -32,13 +32,13 @@ public class SmallScene {
 	public double getLatitude() {
 		return latitude;
 	}
-	public void setLatitude(int latitude) {
+	public void setLatitude(double latitude) {
 		this.latitude = latitude;
 	}
 	public double getLongtitude() {
 		return longtitude;
 	}
-	public void setLongtitude(int longtitude) {
+	public void setLongtitude(double longtitude) {
 		this.longtitude = longtitude;
 	}
 	public int getBigSceneID() {
